+++ react-practice-two/src/pages/Shelf/MyShelf.jsx
@@ -5,10 +5,9 @@ import { useToast } from '../../components/Toast/ToastProvider';
 import { getCurrentUserId } from '../../services/servicesUser';
 import { getListBookById } from '../../services/servicesBook';
 import React, { useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import HeartIcon from '../../components/Icon';
-import Button from '../../components/Button';
 import Header from '../../layouts/Header';
+import BorrowBooks from '../BorrowBooks';
+import FavoriteBooks from '../FavoriteBooks';
 
 const MyShelf = () => {
     const [filteredFavorites, setFilteredFavorites] = useState([]);
@@ -17,7 +16,6 @@ const MyShelf = () => {
     const { showLoading, hideLoading } = useLoading();
     const [favorites, setFavorites] = useState([]);
     const [books, setBooks] = useState([]);
-    const navigate = useNavigate();
     const addToast = useToast();
 
     useEffect(() => {
@@ -102,70 +100,10 @@ const MyShelf = () => {
                             <button className={`tab ${currentTab === 'favorites' ? 'active' : ''}`} onClick={() => handleTabChange('favorites')}>Favorite</button>
                         </div>
                         {currentTab === 'all' && (
-                            <div className="books-list">
-                                {filteredBooks.map(book => (
-                                    <article key={book._id.$oid} className="book-item">
-                                        <div className='book-item-column-left'>
-                                            <img src={book.urlImage} alt={book.name} className="book-item-image" />
-                                            <h3 className='book-item-name'>{book.name}</h3>
-                                            <p className='book-item-author'>{book.author}</p>
-                                            <p className="book-item-rate">5.0 Ratings</p>
-                                        </div>
-                                        <aside className="book-item-column-right">
-                                            <div className="book-item-column-right-text">
-                                                <p className='book-item-status'>Borrowed on </p>
-                                                <p className='book-item-time'>11 Mar 2023 09:00 AM</p>
-                                            </div>
-                                            <Button
-                                                onClick={() => handleReturnBook(book.shelfId)}
-                                                text="Return"
-                                                className="btn-return"
-                                                size="btn-large"
-                                            />
-                                        </aside>
-                                    </article>
-                                ))}
-                            </div>
+                            <BorrowBooks books={filteredBooks} handleReturnBook={handleReturnBook} />
                         )}
                         {currentTab === 'favorites' && (
-                            <div className="books-list">
-                                <div className='wrap-book-item-favourite-title'>
-                                    <p className='title-one'>Title</p>
-                                    <p className='title-two'>Ratings</p>
-                                    <p className='title-three'>Category</p>
-                                    <p className='title-four'>Status</p>
-                                </div>
-                                {filteredFavorites.map(book => (
-                                    <div key={book._id.$oid} className="book-item-favourite">
-                                        <img src={book.urlImage} alt={book.name} className="book-item-favourite-image" />
-                                        <div className='book-item-favourite-content'>
-                                            <h3 className='book-item-name'>{book.name}</h3>
-                                            <p className='book-item-author'>{book.author}</p>
-                                        </div>
-                                        <p className="book-item-favourite-rate">4.5/<span>5</span></p>
-                                        <div className='book-item-favourite-category'>
-                                            <p>Computer Science</p>
-                                            <p>Ux Design</p>
-                                        </div>
-                                        <Button
-                                            text="In-Shelf"
-                                            className="btn-In-Shelf"
-                                            color="btn-enable"
-                                            borderRadius="btn-rounded"
-                                        />
-                                        <HeartIcon
-                                            className="heart-icon"
-                                            onClick={() => handleUnlikeBook(book.favoriteId)}
-                                        />
-                                        <Button
-                                            onClick={() => navigate(`/preview-page/${book._id.$oid}`)}
-                                            text="Preview"
-                                            className="btn-preview"
-                                            borderRadius="btn-rounded"
-                                        />
-                                    </div>
-                                ))}
-                            </div>
+                            <FavoriteBooks favorites={filteredFavorites} handleUnlikeBook={handleUnlikeBook} />
                         )}
                     </div>
                 </div>
@@ -175,4 +113,3 @@ const MyShelf = () => {
 };
 
 export default MyShelf;
-
